create private module orientation._lib
